Biogas-Market - Forecast(2025 - 2031)
The global biogas market size was valued at USD 44.5 billion in 2022 and is projected to reach USD 65.57 billion by 2030, at a CAGR of 5.3% during the forecast period. Europe market has captured revenue share of around 42.5% in 2022. The growth of the biogas market can be attributed to the increasing demand for renewable energy sources, rising energy costs, and government support for sustainable energy development.

𝐆𝐞𝐭 𝐭𝐡𝐞 𝐅𝐮𝐥𝐥 𝐑𝐞𝐩𝐨𝐫𝐭 @ https://tinyurl.com/mvvpmjtr